requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
python-dotenv
firebase-admin
python-multipart